+++ experimental/dds/tree2/src/feature-libraries/chunked-forest/codec/schemaBasedEncoding.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { assert, unreachableCase } from "@fluidframework/core-utils";
+import { unreachableCase } from "@fluidframework/core-utils";
 import {
 	FieldStoredSchema,
 	ITreeCursorSynchronous,
@@ -14,8 +14,7 @@ import {
 import { FieldKind, FullSchemaPolicy, Multiplicity } from "../../modular-schema";
 import { fail } from "../../../util";
 import { fieldKinds } from "../../default-field-kinds";
-import { ICodecOptions, IJsonCodec } from "../../../codec";
-import { EncodedChunk, EncodedValueShape, Versioned, validVersions } from "./format";
+import { EncodedChunk, EncodedValueShape } from "./format";
 import {
 	EncoderCache,
 	FieldEncoder,
@@ -27,7 +26,6 @@ import {
 	compressedEncode,
 } from "./compressedEncode";
 import { NodeShape } from "./nodeShape";
-import { decode } from "./chunkDecoding";
 
 /**
  * Encode data from `cursor` in into an `EncodedChunk`.
@@ -131,34 +129,4 @@ function valueShapeFromSchema(schema: ValueSchema | undefined): undefined | Enco
 		default:
 			unreachableCase(schema);
 	}
-}
-
-export function makeSchemaCompressedCodec(
-	{ jsonValidator: validator }: ICodecOptions,
-	schema: SchemaData,
-	policy: FullSchemaPolicy,
-): IJsonCodec<ITreeCursorSynchronous, string> {
-	const versionedValidator = validator.compile(Versioned);
-	const formatValidator = validator.compile(EncodedChunk);
-	return {
-		encode: (data: ITreeCursorSynchronous) => {
-			const encoded = schemaCompressedEncode(schema, policy, data);
-			assert(versionedValidator.check(encoded), "Encoded schema should be versioned");
-			assert(formatValidator.check(encoded), "Encoded schema should validate");
-			return JSON.stringify(encoded);
-		},
-		decode: (data: string): ITreeCursorSynchronous => {
-			const parsed = JSON.parse(data);
-			if (!versionedValidator.check(parsed)) {
-				fail("invalid serialized schema: did not have a version");
-			}
-			if (!formatValidator.check(parsed)) {
-				if (validVersions.has(parsed.version)) {
-					fail("Unexpected version for schema");
-				}
-				fail("Serialized schema failed validation");
-			}
-			return decode(parsed).cursor();
-		},
-	};
 }
