+++ experimental/dds/tree2/src/feature-libraries/chunked-forest/codec/uncompressedEncode.ts
@@ -3,21 +3,10 @@
  * Licensed under the MIT License.
  */
 
-import { assert } from "@fluidframework/core-utils";
-import { ICodecOptions, IJsonCodec } from "../../../codec";
 import { ITreeCursorSynchronous, forEachField, forEachNode } from "../../../core";
 import { FluidSerializableReadOnly } from "../../contextuallyTyped";
-import { fail } from "../../../util";
-import {
-	EncodedChunk,
-	version,
-	EncodedTreeShape,
-	EncodedNestedArray,
-	Versioned,
-	validVersions,
-} from "./format";
+import { EncodedChunk, version, EncodedTreeShape, EncodedNestedArray } from "./format";
 import { ShapeIndex } from "./formatGeneric";
-import { decode } from "./chunkDecoding";
 
 /**
  * Encode data from `cursor` in the simplest way supported by `EncodedChunk`.
@@ -64,32 +53,4 @@ function encodeSequence(cursor: ITreeCursorSynchronous): FluidSerializableReadOn
 		data.push(local);
 	});
 	return data;
-}
-
-export function makeUncompressedCodec({
-	jsonValidator: validator,
-}: ICodecOptions): IJsonCodec<ITreeCursorSynchronous, string> {
-	const versionedValidator = validator.compile(Versioned);
-	const formatValidator = validator.compile(EncodedChunk);
-	return {
-		encode: (data: ITreeCursorSynchronous) => {
-			const encoded = uncompressedEncode(data);
-			assert(versionedValidator.check(encoded), "Encoded schema should be versioned");
-			assert(formatValidator.check(encoded), "Encoded schema should validate");
-			return JSON.stringify(encoded);
-		},
-		decode: (data: string): ITreeCursorSynchronous => {
-			const parsed = JSON.parse(data);
-			if (!versionedValidator.check(parsed)) {
-				fail("invalid serialized schema: did not have a version");
-			}
-			if (!formatValidator.check(parsed)) {
-				if (validVersions.has(parsed.version)) {
-					fail("Unexpected version for schema");
-				}
-				fail("Serialized schema failed validation");
-			}
-			return decode(parsed).cursor();
-		},
-	};
 }
